WIP. refactoring StructureFromMotion. Added some synthetic scenes for testing spatial pipeline components (gtsam stuff). Feature manager implementation and incorporation in StructureFromMotion still needs work.
